Separate API handler code and rename imports
